requirements.txt
python-dotenv==1.0.1
tortoise-orm==0.21.5
asyncpg==0.29.0
aiogram==3.10.0
beautifulsoup4==4.12.3
aiohttp==3.9.5